Fix: Previous key state for repeat count in X11
Currently, as we've spoken in https://github.com/aseprite/laf/pull/69
@dcap has correctly pointed out that my patch has some problems with
shift key modifier - and he was correct, I've skipped an important thing
looking into this problem.

This patch adds a std::set to hold all currently pressed keys - if a key
got pressed second time, std::set will hold a key symbol of that key, if
an event for that key comes for a second time (and there were no
KeyRelease event coming for the event in-between) - then it means the key is
being pressed down, so we set repeat count to 1. If KeyRelease event
came for the key symbol that was previously pressed, then we basically
erase it from std::set.