Fix base::get_absolute_path() fixing base::normalize_path() function
base::get_absolute_path() resulted in bugs when more than two or more
double dots were present in the input path.
The source of the error was the base::normalize_path() function
which has been fixed.
Also added tests in 'fs_tests.cpp'.
This fix is part of the solution of aseprite/aseprite#4851.